third_party/WebKit/LayoutTests/inspector-protocol/debugger/debugger-doesnt-step-into-injected-script.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script>

function foo()
{
    return 239;
}

function test()
{
    InspectorTest.sendCommandOrDie("Debugger.enable", {});
    InspectorTest.eventHandler["Debugger.paused"] = debuggerPaused;
    InspectorTest.sendCommandOrDie("Runtime.evaluate", { "expression": "(function boo() { setTimeout(foo, 0); debugger; })()" });

    var actions = [ "stepInto", "stepInto", "stepInto" ];
    function debuggerPaused(result)
    {
        InspectorTest.log("Stack trace:");
        for (var callFrame of result.params.callFrames)
            InspectorTest.log(callFrame.functionName + ':' + callFrame.location.lineNumber + ":" + callFrame.location.columnNumber);
        InspectorTest.log("");

        var action = actions.shift();
        if (!action) {
            InspectorTest.sendCommandOrDie("Debugger.resume", {}, () => InspectorTest.completeTest());
            return;
        }
        InspectorTest.log("Perform " + action);
        InspectorTest.sendCommandOrDie("Debugger." + action, {});
    }
}
</script>
</head>
<body onLoad="runTest();">
Check that stepInto at then end of the script go to next user script instead InjectedScriptSource.js.
</body>
</html>
