third_party/blink/web_tests/inspector-protocol/debugger/debugger-doesnt-step-into-injected-script-expected.txt

Check that stepInto at then end of the script go to next user script instead InjectedScriptSource.js.
Stack trace:
boo:0:38
:0:50

Perform stepInto
Stack trace:
boo:0:48
:0:50

Perform stepInto
Stack trace:
:0:52

Perform stepInto
Stack trace:
foo:7:4
